services/sync-worker/pyproject.toml
[project]
name = "sync-worker"
version = "0.1.0"
description = "Tidyboard CalDAV sync worker"
license = { text = "AGPL-3.0-or-later" }
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "caldav>=3.1.0",
    "icalendar>=5.0.0",
    "python-dateutil>=2.9.0",
    "structlog>=24.1.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "vcrpy>=6.0.0",
    "testcontainers>=4.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sync_worker"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-p no:parallel -x -v"
asyncio_mode = "auto"
pythonpath = ["src"]
markers = [
    "integration: marks tests as integration tests (require network or external services)",
]

[tool.ruff]
line-length = 100
target-version = "py311"
